requirements.txt
Flask
flask-apscheduler
influxdb-client
pyyaml
requests
